Added math_msvc_compatibility.h
MSVC++ 9 has a very outdated definition of <math.h>.
That is why math_msvc_compatibility.h should be imported
(if and only if _MSC_VER is defined).

M_PI is only defined by <math.h> when _USE_MATH_DEFINES
is defined.

isnan, isinf, isfinite are not defined in <math.h>.
Instead, _isnan and _finite must be used.

Furthermore, <math.h> defines copysign as a macro, which
interferes with some macro-heavy code in numpy_quaternion.c.
This is why the wrapper functions are defined as static
inline functions rather than macros.