Add INCREFs to inplace operators [closes #20]
Apparently, the inplace operators decref the return value, rather than
leaving it alone.  I would expect a line like

    a *= b

to have no effect on the reference count of `a`.  Evidently, I was
wrong.  The function would return just fine, and the first use of `a`
would be fine because the stale memory was still being found and used.
But that typically overwrites the memory with a string or something,
so later uses would go haywire.

I've added Py_INCREF to each return from the inplace C
functions, and checked that there is no memory leak using `mprof`.