Make slerp- and squad-associated functions more uniform
Previously, there was no particular consistency in the names or usage of the various slerp and squad functions.  There are now six functions:

    squad_evaluate      slerp_evaluate
    squad_vectorized    slerp_vectorized
    squad               slerp

The `evaluate` functions just evaluate one input; the `vectorized` functions loop efficiently at the C level; and the plain functions are simplest to use, and adjust the times to be relative.

Closes #28