Fix comparisons in some tests.
test_quaternion_absolute() and test_quaternion_norm(): use np.allclose
for comparison to avoid failing with a small floating-point error.

test_as_spherical_coords(): the input values for varphi are on [0, 2pi]
but the returned values are calculated with np.arctan2 and are on [-pi,
pi]. Add 2*pi to negative return values to allow comparison.